Store the settings into the DB
Write the service settings into the DB. If the service is not yet in the DB,
add a record for that, too.

In order to achieve this, some functions have been added to the SQL, to set and
get the row ID of the newly creted account/service records.

When a new service record is written, the ID is also set in the AgService
structure in the AgManager cache.