Give each Service an ID when loading it
The late creation of services into the DB was causing problems when a process
had more than one Manager instance running: if one instance added a service to
the DB, the second one wouldn't notice it.

This patch changes the situation by storing the services into the DB as soon as
they are loading.
As a result, some code to obtain/update the service ID after creation could be
removed.